iterate refactor
