+++ library/react-polkicon/src/types.ts
@@ -1,0 +1,14 @@
+/* @license Copyright 2024 w3ux authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+export interface Circle {
+  cx: number;
+  cy: number;
+  fill: string;
+  r: number;
+}
+
+export interface Scheme {
+  freq: number;
+  colors: number[];
+}
