+++ library/react-polkicon/src/utils.tsx
@@ -1,32 +1,13 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import { decodeAddress, blake2AsU8a } from "@w3ux/utils/util-crypto";
+import { AccountId } from "@polkadot-api/substrate-bindings";
+import { blake2AsU8a } from "@polkadot/util-crypto";
+import { Scheme } from "./types";
+import { PolkiconCenter, SCHEMA } from "./consts";
 
-/* A generic identity icon, taken from:
+/* A generic identity icon. Circle generation logic taken from:
 https://github.com/polkadot-js/ui/tree/master/packages/react-identicon */
-
-export interface Circle {
-  cx: number;
-  cy: number;
-  fill: string;
-  r: number;
-}
-
-const SIZE = 64;
-const C = SIZE / 2;
-export const Z = (SIZE / 64) * 5;
-
-export const outerCircle = (fill: string): Circle => ({
-  cx: C,
-  cy: C,
-  fill,
-  r: C,
-});
-
-export const renderCircle = ({ cx, cy, fill, r }: Circle, key: number) => (
-  <circle cx={cx} cy={cy} fill={fill} key={key} r={r} />
-);
 
 // Calculates and returns several rotational values based on an initial parameter C.
 //
@@ -46,7 +27,7 @@ const getRotation = (): {
   const param_ro4 = 4;
   const param_r3o4 = 4;
 
-  const r = (C / 4) * param_r;
+  const r = (PolkiconCenter / 4) * param_r;
   const rroot3o2 = (r * Math.sqrt(3)) / param_rroot3o2;
   const ro2 = r / param_ro2;
   const rroot3o4 = (r * Math.sqrt(3)) / param_rroot3o4;
@@ -59,6 +40,10 @@ const getRotation = (): {
 export const getCircleCoordinates = (): [number, number][] => {
   const { r, r3o4, ro2, ro4, rroot3o2, rroot3o4 } = getRotation();
 
+  // Alias center coordinate for more concise code.
+  const C = PolkiconCenter;
+
+  // Return all coordinates
   return [
     [C, C - r],
     [C, C - ro2],
@@ -82,42 +67,6 @@ export const getCircleCoordinates = (): [number, number][] => {
   ];
 };
 
-interface Scheme {
-  freq: number;
-  colors: number[];
-}
-
-export const SCHEMA: Record<string, Scheme> = {
-  target: {
-    colors: [0, 28, 0, 0, 28, 0, 0, 28, 0, 0, 28, 0, 0, 28, 0, 0, 28, 0, 1],
-    freq: 1,
-  },
-  cube: {
-    colors: [0, 1, 3, 2, 4, 3, 0, 1, 3, 2, 4, 3, 0, 1, 3, 2, 4, 3, 5],
-    freq: 20,
-  },
-  quazar: {
-    colors: [1, 2, 3, 1, 2, 4, 5, 5, 4, 1, 2, 3, 1, 2, 4, 5, 5, 4, 0],
-    freq: 16,
-  },
-  flower: {
-    colors: [0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 3],
-    freq: 32,
-  },
-  cyclic: {
-    colors: [0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5, 6],
-    freq: 32,
-  },
-  vmirror: {
-    colors: [0, 1, 2, 3, 4, 5, 3, 4, 2, 0, 1, 6, 7, 8, 9, 7, 8, 6, 10],
-    freq: 128,
-  },
-  hmirror: {
-    colors: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 8, 6, 7, 5, 3, 4, 2, 11],
-    freq: 128,
-  },
-};
-
 // Takes a number `d` and returns a scheme from `SCHEMA` based on cumulative frequency.
 //
 // Iterates over each scheme and checks if the cumulative frequency exceeds `d`. If `d` is less than
@@ -128,13 +77,10 @@ export const SCHEMA: Record<string, Scheme> = {
 export const findScheme = (d: number): Scheme =>
   Object.values(SCHEMA).find((scheme) => (d -= scheme.freq) < 0);
 
-let zeroHash: Uint8Array = new Uint8Array();
-
 const addressToId = (address: string): Uint8Array => {
-  if (!zeroHash.length) {
-    zeroHash = blake2AsU8a(new Uint8Array(32));
-  }
-  const pubKey = decodeAddress(address || "").toString();
+  const zeroHash = blake2AsU8a(new Uint8Array(32));
+  const codec = AccountId();
+  const pubKey = codec.dec(codec.enc(address)).toString();
   return blake2AsU8a(pubKey).map(
     (x: number, i: string | number) => (x + 256 - zeroHash[i]) % 256
   );
